+++ src/lib.rs
@@ -56,7 +56,7 @@ pub use buffered::{
     BufferedSource, BufferedSourceBuilder, DEFAULT_BLOCK, DEFAULT_LOOKBACK_DEN,
     DEFAULT_LOOKBACK_NUM, DEFAULT_PREFETCH_TIMEOUT,
 };
-pub use concat::open_concat;
+pub use concat::{open_concat, parse as parse_concat_uri, ConcatUri};
 pub use data::{open_data, parse as parse_data_uri, DataUri};
 pub use file::open_file;
 pub use mem::open_mem;